src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
DateTime==4.3
Deprecated==1.2.13
Flask==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.41.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
json-logging==1.3.0
logger==1.4
MarkupSafe==2.0.1
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-exporter-jaeger-proto-grpc==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-exporter-zipkin==1.7.1
opentelemetry-exporter-zipkin-json==1.7.1
opentelemetry-exporter-zipkin-proto-http==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-flask==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.5
protobuf==3.19.1
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
thrift==0.15.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zope.interface==5.4.0
